Some fixef after refactoring
