+++ src/components/burger-ingredients/burger-ingredients.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { TStore } from "../../../declarations/store";
+import { TStore } from "../../declarations/store";
 import curStyle from './burger-ingredients.module.css';
 import { Tab } from '@ya.praktikum/react-developer-burger-ui-components';
 import Ingredient from '../ingredient/ingredient';
