+++ src/components/protected-route-element/protected-route-element.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import { Navigate, useLocation } from 'react-router-dom';
-import { TStore } from '../../../declarations/store';
+import { TStore } from '../../declarations/store';
 
 interface IProtectedRouteElement {
   children: React.ReactNode;
