+++ src/declarations/store.ts
@@ -1,11 +1,11 @@
-import { TBurgerConstructorState } from '../src/services/reducers/burger-constructor';
-import { TAuthState } from '../src/services/reducers/auth';
-import { TIgredientsState } from '../src/services/reducers/ingredient';
-import { TIngredientDetailsState } from '../src/services/reducers/ingredient-details';
-import { TOrderState } from '../src/services/reducers/order';
-import { TForgotPasswordState } from '../src/services/reducers/forgot-password';
-import { TResetPasswordState } from '../src/services/reducers/reset-password';
-import { TProfileState } from '../src/services/reducers/profile';
+import { TBurgerConstructorState } from '../services/reducers/burger-constructor';
+import { TAuthState } from '../services/reducers/auth';
+import { TIgredientsState } from '../services/reducers/ingredient';
+import { TIngredientDetailsState } from '../services/reducers/ingredient-details';
+import { TOrderState } from '../services/reducers/order';
+import { TForgotPasswordState } from '../services/reducers/forgot-password';
+import { TResetPasswordState } from '../services/reducers/reset-password';
+import { TProfileState } from '../services/reducers/profile';
 
 export type TStore = {
   burger: TBurgerConstructorState;
